feat(ocean/aws): add missing fields to ResourceSuggestion type
The API also exposes the fields `resourceType` and `resourceName` but
these are currently not part of the `ResourceSuggestion` type.

This makes resource suggestions for DaemonSets and StatefulSets not very
useful as the information about the resource are missing.

This change adds the missing fields `ResourceName` and `ResourceType`,
and while we're at it also adds the missing `Containers` field.